Some Waveshaper Fixes (#396)
- Rip out all the old filter nonsense we dont need
- Fix the drive range properly
- Add Biquads for LP/HP
- Fix the double-click-to-reset mystery UI thing

+++ src/Waveshaper.cpp
@@ -50,10 +50,10 @@ struct WaveshaperSelector : widgets::ParamJogSelector
         int type = (int)std::round(getParamQuantity()->getValue());
         auto di = wsm.remapStreamedIndexToDisplayIndex(type);
         di += dir;
-        if (di >= wsmOrdering.size())
+        if (di >= (int)wsmOrdering.size())
             di = 0;
         if (di < 0)
-            di = wsmOrdering.size() - 1;
+            di = (int)wsmOrdering.size() - 1;
         setType(wsmOrdering[di]);
     }
